splitted connection models
