+++ strongMan/apps/connections/models/__init__.py
@@ -1,0 +1,3 @@
+from .connections import *
+from .authentication import *
+from .specific import *
