+++ strongMan/apps/connections/request_handler/DeleteHandler.py
@@ -5,7 +5,7 @@ from django.core.urlresolvers import reverse
 from strongMan.apps.vici.wrapper.exception import ViciExceptoin
 from strongMan.apps.vici.wrapper.wrapper import ViciWrapper
 
-from ..models import Connection
+from strongMan.apps.connections.models.connections import Connection
 
 
 class DeleteHandler:
