+++ strongMan/integration_tests/test_integration.py
@@ -3,7 +3,8 @@ import os
 from django.test import TestCase, Client
 from django.contrib.auth.models import User
 
-from strongMan.apps.connections.models import Connection, Child
+from strongMan.apps.connections.models.specific import Child
+from strongMan.apps.connections.models.connections import Connection
 from strongMan.apps.vici.wrapper.wrapper import ViciWrapper
 from strongMan.apps.certificates.container_reader import X509Reader, PKCS1Reader
 from strongMan.apps.certificates.services import UserCertificateManager
